# swagger.json
{
    "swagger": "2.0",
    "info": {
      "version": "1.0.0",
      "title": "Askito Rest api",
      "description": "A community for developers to share their knowledge"
    },
    "host": "askito.herokuapp.com",
    "basePath": "/api/",
...
    "schemes": ["https"],
    "consumes": ["application/json"],
    "produces": ["application/json"],
    "paths": {
...
      "/answers/add": {
        "post": {
          "tags": ["Answers"],
          "summary": "add new Answer",
          "description": "",
          "operationId": "addAnswer",
          "consumes": [
            "application/json"
          ],
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "in": "header",
              "name": "x-auth-token",
              "description": "token to be verified",
              "required": true,
              "type": "string"
            },
            {
              "in": "body",
              "name": "body",
              "description": "Answer object that needs to be added to the system",
              "required": true,
              "schema": {
              "$ref": "#/definitions/Answer"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Answer updated successfully"
            },
            "400": {
              "description": "Invalid user"
            },
            "500": {
              "description": "something went wrong try after some time!"
            }
          }
        }
      },
      "/answers/update/{id}": {
        "post": {
          "tags": ["Answers"],
          "summary": "Update Answer",
          "description": "",
          "operationId": "UpdateAnswer",
          "consumes": [
            "application/json"
          ],
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "in": "header",
              "name": "x-auth-token",
              "description": "token to be verified",
              "required": true,
              "type": "string"
            },
            {
              "in": "body",
              "name": "body",
              "description": "Answer object that needs to be added to the system",
              "required": true,
              "schema": {
              "$ref": "#/definitions/Answer"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Answer updated successfully"
            },
            "400": {
              "description": "Invalid user"
            },
            "500": {
              "description": "something went wrong try after some time!"
            }
          }
        }
      },
...
      "/notifications/get/": {
        "get": {
          "tags": ["Notifications"],
          "summary": "Get Notifications of the loged in user",
          "description": "",
          "operationId": "getNotificationsOfLogedInUser",
          "consumes": [
            "application/json"
          ],
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "in": "header",
              "name": "x-auth-token",
              "description": "token to be verified",
              "required": true,
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "",
              "schema": {
                "$ref": "#/definitions/Notification"
              }
            },
            "404": {
              "description": "no notifications found"
            },
            "500": {
              "description": "something went wrong try after some time!"
            }
          }
        }
      },
      "/notifications/update/noticed/{id}": {
        "put": {
          "tags": ["Notifications"],
          "summary": "update Notification noticed",
          "description": "",
          "operationId": "updateNotificationNoticed",
          "consumes": [
            "application/json"
          ],
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "in": "header",
              "name": "x-auth-token",
              "description": "token to be verified",
              "required": true,
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "updated successfully"
            },
            "400": {
              "description": "Invalid user"
            },
            "500": {
              "description": "something went wrong try after some time!"
            }
          }
        }
      },
      "/comments/add": {
        "post": {
          "tags": ["Comments"],
          "summary": "add new comment",
          "description": "",
          "operationId": "addComment",
          "consumes": [
            "application/json"
          ],
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "in": "header",
              "name": "x-auth-token",
              "description": "token to be verified",
              "required": true,
              "type": "string"
            },
            {
              "in": "body",
              "name": "body",
              "description": "types are answer,question,post",
              "required": true,
              "schema": {
              "$ref": "#/definitions/Comment"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "comment updated successfully"
            },
            "400": {
              "description": "Invalid user"
            },
            "500": {
              "description": "something went wrong try after some time!"
            }
          }
        }
      },
      "/comments/update/{id}": {
        "put": {
          "tags": ["Comments"],
          "summary": "update comment",
          "description": "",
          "operationId": "updateComment",
          "consumes": [
            "application/json"
          ],
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "in": "header",
              "name": "x-auth-token",
              "description": "token to be verified",
              "required": true,
              "type": "string"
            },
            {
              "in": "body",
              "name": "body",
              "description": "types are answer,question,post",
              "required": true,
              "schema": {
              "$ref": "#/definitions/Comment"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "comment updated successfully"
            },
            "400": {
              "description": "Invalid user"
            },
            "500": {
              "description": "something went wrong try after some time!"
            }
          }
        }
      },
...
    },
      "definitions": {
...
        "Answer": {
          "type": "object",
          "properties": {
            "answer": {
              "type": "string"
            },
            "questionId" : {
              "type" : "string"
            }
          }
        },
...
        "Notification": {
          "type": "object",
          "properties": {
            "notification": {
              "type": "string"
            }
          }
        },
        "Comment": {
          
          "type": "object",
          "properties": {
            "comment": {
              "type": "string"
            },
            "postId" : {
              "type" : "string"
            },
            "type" :{
              "type" : "answer | question | post"
            }
          }
        },
...
      } 
}